Move long tests in test_ham_restricted_chemical
